Shrink the built distribution size
The "prod" deployment was nearly 1mb combined. Limiting the ngb import
to just the collapse module makes a small dent, but most of the unused
weight seems to be coming from firebase. Things will probably get bigger
once we start adding that back in, but might as well keep things smaller
while we can.