Drop the staging submodule
This wasn't working out so well and the extra commits for deployments
were getting messy. The whole point of maintaining a separate repo was
to prevent the size of our source code repo from getting out of hand,
but pointing to the root folder of a repo as our output folder means
that we delete the whole folder (including git meta files) before
rebuilding which is a pain. Maybe in the future I'll start building into
a docs folder of a separate repo and publish from the docs folder. We'll
see.