Implement ReSTful navigation of choices
I'm still struggling with how (and whether to) make choice urls restful,
but I should probably write a blog post rather than a commit message
about that.

While we had a basic view that worked and displayed random images on
initial page load, subsequent choice requests really only worked the
same way -- by making brand new requests to the home page. This doesn't
help with a few things -- 1) when users hit the back button, we'd like
them to see the last set of images that they saw before they clicked on
one and 2) we want to keep track of a users selections for asyncronous
remote batch logging (in other words when the user figures out what they
want to eat, we'd like to send back a list of all the individual choices
that helped them make that decision).

Restful choice routes seem like one conventional way to handle some of
this -- if the route (url) itself says what foods we should be offering
up in the choice and we just show what the user "requested", then back
navigation will just work out of the box (to a point). However, this
means that we always have to be looking one step ahead -- on two
branches -- which means that rather than coming up with one set of foods
to propose for each request, we have to come up with two -- one for each
possible subsequent request -- and we have to build a URL for each of
the food links that will spawn that request.

I've done just that, but in a really ugly way where the view completely
bled its way into the model. I think it's about time to really separate
out the food model from the choice model -- but that will probably have
to wait a little longer.

I'm not sure that we're really hooking in to the event lifecycle at all
the right points -- this implementation has us redundantly calling the
"getChoice" method (basically rebuilding the model) at three separate
points:

1. on init
2. on click
3. "NavigationEnd" -- an arbitrary router event -- not sure it's really
the most appropriate point in the router lifecycle to hook into

It's important to unsubscribe/resubscribe to that last one only for the
component specific logic, otherwise it runs on all page views, not just
choice page views. I did some random clicking and the method only seems
to get called once per click, not sure why 2 and 3 don't both end up
firing the method, it just seems to work.