Set global style for main padding
The header component is using absolute positioning with a z index so
that it floats over the rest of the body. That way it doesn't effect any
of the page positioning if and when you expand the drop down menu, but
rather it obscures it. When the drop down is collapsed, we don't want to
obscure any of the pages main body, so we're setting a global padding to
that main element that's the size of the header.

I already had some padding set in the food canvas element of the choice
component, but we don't want to have to do this separately for every
page so I'm extracting it out the the global css. There's probably a
better way to do it, but this is good enough for now.