Fix browser back button redirect infinite loops
When a user clicks on the home or choice page, they're automatically
redirected to a page url with the images to display. The original page
gets pushed onto the history stack though in addition to the page that
the user is redirected to, meaning that if they click on the back
button, they go back to the page that will redirect them to their
current page (rather than going back to the last page that they manually
navigated from). To avoid this behavior, we're simply replacing the
original page that got added to the history with the one that we
redirect the user to.