Refactor -- extract choice out of food model
This is still a bit messy, but at least now the choice service is
responsible for providing "choices" -- or two foods to pair off together
-- and the food service is simply responsible for providing foods, not
coming up with which foods to put up side by side in a "food fight".
Similarly, neither is responsible for building the routes or urls for a
given choice, or the choice that is represented in the url that will be
routed to when a use clicks on a food image. Now that is purely a view
side (choice component) modeling decision.