Add a counter to choice routes
In the real world this might not really be necessary, but I couldn't
help noticing in our mock food world where we only have two different
food items and half of the time a link to the next random choice will be
to the same exact choice url that the router navigation events don't
happen somebody clicks on a link to the same exact page. A click event
happens, so we can react to the event, however it doesn't update the
browser history, so suddenly our event count is out of sync with the
navigation event count.

Adding a step counter into the next choice link urls forces clicks to
always update the router. It also gives us some a simple and convenient
counter for seeing how far down the choice tree the user has clicked and
depending on how we track the clicks it will also show us if they hit
the back button signifying that maybe the made a mistake and we should
ignore the last click they made.

Again, this may all get stripped out in the future, but it seems to make
some sense for now.