Add some more food items
We're about to start implementing the "crystal food bowl" featuer which
will try to do a better job of guessing what you're hungry for based on
what images you click using a binary search approach. It depends on a
decision tree of food items clustered by similarity and it will be hard
to test out jusing just two images so I'm adding in a few more before
implementing the logic.