Implement binary food search model
This is just a first stab at our ultimate goal of a "crystal ball" (or
crystal bowl) for magically predicting what users are hungry for.

For now I've just dropped it in place of the "food fight" choice model
that's been running under various implementations for the last year
showing users random food stand offs.

We plan to continue hosting both interfaces side by side (along with
possible others) and allowing users to choose between them with one
being the default home page route.

The easiest (and ungliest) way to get something up and running as a
quick demo though was to drop the new implementation right in place of
the old one. For now it just loops back to the start if you reach a
terminal node -- eventually it should stop and ask if that's what you
want and route you to follow on functionality, but that's all yet to be
determined.

It's very messy and needs a lot of refactoring, but the rough
implementation has been informative in helping to figure out what parts
of the api to extract into a more generic interface.