Implement basic consent for anonymous data collection
We want to make sure that we have explicit consent from users before
sending their appetite (food click) history silently in the background.
This is just a start -- still need to refine the UI design and wording
itself as well as store the users response in local storage so that we
don't have to ask them again everytime they come back to the site.