Use plain old local storage
I started off with the @ngx-pwa/local-storage async local storage
framework, but I'm not sure that it was really adding any huge
advantages at this point. Eventually it may be a better decision to use
it as it wraps the indexedDB webstorage facility which may have some
performance and security advantages, but for storing the users decision
of whether or not to let us collect their food clicks anonymously, and
even storing a short history of innocuous food clicks, there doesn't
seem to be any big need for the extra complexity at this point.