Refactor appetite service to use local storage
Up until now we were losing appetite information if the user left the
site (for instance if they clicked on an attribution link). While this
meant that we couldn't use those details to help provide a better food
recommendation (which hasn't really been an issue so far), it also meant
that for users that were happy to send us their anonymous appetite
preferences in the background, we would often lose the clicks before we
got around to sending the history (if for instance, say the user were to
close the tab before idling long enough). Now we may still lose some
click details if the user closes the tab and never comes back to the
site, but for frequent users who are happy to send us their food
preferences, even if they don't send them during their session, they'll
usually still be in the browsers local storage the next time the user
comes back to the site and we'll send them then (as long as the user
doesn't clear out their storage or use private browsing mode...)