Refactor for google cloud/firebase hosting
As part of switching over to google cloud for our hosting, we're also
moving all of our images and other static assets over to google's cloud
storage -- partly to manage cors headers, partly to separate staged
assets from produciton ones, and partly to position ourselves better for
the future of user generated content (users uploading and annotating
their own food images). This meant that we had to move the location of
the base food url, but we also decided to restructure some of the paths
within that base url context -- mostly for easier management of assets.