Implement basic firebase authentication
This is still a pretty rough work in progress, but for now it will be
running in stealth mode on the staging server only, so it doesn't need
to look pretty because we'll only be using it for admins to clean up
some of our metadata. We'll come back and make it look prettier after
we've moved the metadata over to firestore and had some practical
experience actually using it.