Implement basic CRUD support for food photos
This is just a rough initial implementation of firebase backed photo
metadata support. I still need to make it look nicer, refactor a lot of
the redundancy into separate components and add in drag and drop support
for moving tags between tag lists, but wanted to check it in as is
before going through a major refactoring effort.