Add httpclient alternative to firestore for dev
Relying on the AngularFire Firestore backend for CRUD operations on the
photo api was making it difficult to work on the actual form development
locally without an internet connection -- it also meant that I needed a
separate backend if i didn't want to constantly be changing data on a
real environment during development.

I'm still struggling with the right way to do this with Angular
dependency injection -- the current fix statically binds all of the
implementations all of the time. Eventually I want to replace this with
a modular approach that lazy loads most of the angularfire modules upon
authentication -- but I still need to figure out how to hot swap one
implementation of a service in place of another one that was already
there.

Basically, this is an ugly hack that kind of works for now, but needs to
get fixed soon.