Add comma as a tag separator
With the ngx-chips tag interface, it may not always be obvious to users
that they need to hit enter to add a new tag. Some users might start
typing in multiple tags separated by a comma -- we don't want tags to
include commas, so adding the comma as an alternative break that starts
a new tag might help alleviate some confusion.