Refactor attributions into dynamic template
As part of the slow migration from the old food api over to the new
photo api, we wanted to get rid of references to the food.attribution
field that came from a separate static html file loaded dynamically with
a separate XHR request. With these changes, the attributions are built
dynamically based on attribution properties in the photo metadata json
file.

We still need to work out the various ways that the photo service might
retrieve those details (static json file hosted with assets vs dynamic
firestore object), but this gets us closer to having the photo model and
photo service manage all of the photo specific details itself paving the
way for our new food api (where the eventual relationship between a food
object and a photo object won't necessarily be one to one).