Implement fallback photo service
Create a photo service that acts as a proxy to two backing services -- a
primary and secondary. The fallback is only used for reads -- writes are
only sent to the primary backing service. This provides a mechanism for
reading lazily from a read only source and updating to a writable cache.

This is a bit of a short term stop-gap until I figure out a better way
to handle moderation workflow patterns that typically involve editing on
one source and then moving to another target (kind of the opposite of
this approach).