Fix auth issue
The auth service hasn't been updating the user document in firestore for
new users (probably since we switched over to using redirects instead of
popups). We didn't notice this until a new user came along and tried
logging in. He was logged in, but couldn't see it because the update
user logic wasn't happening and the visual feedback was based on the
firestore user rather than just using the auth service user object. We
aren't yet at a point where we necessarily want users to have profile
pages that other users can see, so there's no real reason to be saving
the users profile in firestore. All we need for moderation is to copy
over the users uid into the moderator collection.