Fix issue with duplicate images in choices
When we switched the decision tree that Crystal Bowl uses over to a
model that allows redundancy between branches (as in the first few
splits don't separate the nodes completely down the middle, but each
branch includes more than half and consequently nodes from the middle
end up redundantly in both branches), we ended up in a situation where
it was possible that we would select the same image to represent
opposing branches. Not did it become possible, but attempting to choose
foods that only differed by the direction of the branch, and were as
similar as possible on all other dimensions, but rather it actually
became more likely than one would expect if we selected foods randomly
from the branches.

This commit attempts to address the problem with a simple hack that
replaces the food from the second branch with one of it's nearest
neighbors in the food space (if the food matched the food from the first
branch).